+++ sdk/resourcemanager/confidentialledger/armconfidentialledger/models.go
@@ -22,6 +22,21 @@ type AADBasedSecurityPrincipal struct {
 	TenantID *string
 }
 
+// Backup - Object representing Backup properties of a Confidential Ledger Resource.
+type Backup struct {
+	// REQUIRED; SAS URI used to access the backup Fileshare.
+	URI *string
+
+	// The region where the backup of the ledger will eventually be restored to.
+	RestoreRegion *string
+}
+
+// BackupResponse - Object representing the backup response of a Confidential Ledger Resource.
+type BackupResponse struct {
+	// READ-ONLY; Response body stating if the ledger is being backed up.
+	Message *string
+}
+
 // CertBasedSecurityPrincipal - Cert based security principal with Ledger RoleName
 type CertBasedSecurityPrincipal struct {
 	// Public key of the user cert (.pem or .cer)
@@ -76,6 +91,18 @@ type ConfidentialLedger struct {
 	Type *string
 }
 
+// CustomizationData - Object representing the different fields of customization data.
+type CustomizationData struct {
+	// CCF Property for the logging level for the untrusted host: Trace, Debug, Info, Fail, Fatal.
+	HostLevel *string
+
+	// CCF Property for the maximum size of the http request body: 1MB, 5MB, 10MB.
+	MaxBodySize *string
+
+	// CCF Property for the subject name to include in the node certificate. Default: CN=CCF Node.
+	SubjectName *string
+}
+
 // DeploymentType - Object representing DeploymentType for Managed CCF.
 type DeploymentType struct {
 	// Source Uri containing ManagedCCF code
@@ -93,11 +120,17 @@ type LedgerProperties struct {
 	// Array of all cert based Security Principals.
 	CertBasedSecurityPrincipals []*CertBasedSecurityPrincipal
 
+	// SKU associated with the ledger
+	LedgerSKU *LedgerSKU
+
 	// Type of Confidential Ledger
 	LedgerType *LedgerType
 
 	// Object representing RunningState for Ledger.
 	RunningState *RunningState
+
+	// READ-ONLY; Enclave Platform on which Ledger Application is deployed, defaults to IntelSgx
+	EnclavePlatform *EnclavePlatform
 
 	// READ-ONLY; Endpoint for accessing network identity.
 	IdentityServiceURI *string
@@ -148,6 +181,21 @@ type ManagedCCF struct {
 	Type *string
 }
 
+// ManagedCCFBackup - Object representing Backup properties of a Managed CCF Resource.
+type ManagedCCFBackup struct {
+	// REQUIRED; SAS URI used to access the backup Fileshare.
+	URI *string
+
+	// The region where the backup of the managed CCF resource will eventually be restored to.
+	RestoreRegion *string
+}
+
+// ManagedCCFBackupResponse - Object representing the backup response of a Managed CCF Resource.
+type ManagedCCFBackupResponse struct {
+	// READ-ONLY; Response body stating if the managed CCF resource is being backed up.
+	Message *string
+}
+
 // ManagedCCFList - Object that includes an array of Managed CCF and a possible link for next set.
 type ManagedCCFList struct {
 	// The URL the client should use to fetch the next page (per server side paging).
@@ -159,14 +207,23 @@ type ManagedCCFList struct {
 
 // ManagedCCFProperties - Additional Managed CCF properties.
 type ManagedCCFProperties struct {
+	// Additional Managed CCF deployment-time customizations.
+	CustomizationData *CustomizationData
+
 	// Deployment Type of Managed CCF
 	DeploymentType *DeploymentType
+
+	// Enclave Platform on which MCCF Application is deployed, defaults to IntelSgx
+	EnclavePlatform *EnclavePlatform
 
 	// List of member identity certificates for Managed CCF
 	MemberIdentityCertificates []*MemberIdentityCertificate
 
 	// Number of CCF nodes in the Managed CCF.
 	NodeCount *int32
+
+	// Object representing RunningState for Managed CCF.
+	RunningState *RunningState
 
 	// READ-ONLY; Unique name for the Managed CCF.
 	AppName *string
@@ -177,8 +234,26 @@ type ManagedCCFProperties struct {
 	// READ-ONLY; Endpoint for accessing network identity.
 	IdentityServiceURI *string
 
-	// READ-ONLY; Provisioning state of Ledger Resource
+	// READ-ONLY; Provisioning state of Managed CCF Resource
 	ProvisioningState *ProvisioningState
+}
+
+// ManagedCCFRestore - Object representing Restore properties of Managed CCF Resource.
+type ManagedCCFRestore struct {
+	// REQUIRED; Fileshare where the managed CCF resource backup is stored.
+	FileShareName *string
+
+	// REQUIRED; The region the managed CCF resource is being restored to.
+	RestoreRegion *string
+
+	// REQUIRED; SAS URI used to access the backup Fileshare.
+	URI *string
+}
+
+// ManagedCCFRestoreResponse - Object representing the restore response of a Managed CCF Resource.
+type ManagedCCFRestoreResponse struct {
+	// READ-ONLY; Response body stating if the managed CCF resource is being restored.
+	Message *string
 }
 
 // MemberIdentityCertificate - Object representing MemberIdentityCertificate for Managed CCF.
@@ -227,6 +302,24 @@ type ResourceProviderOperationList struct {
 
 	// READ-ONLY; Resource provider operations list.
 	Value []*ResourceProviderOperationDefinition
+}
+
+// Restore - Object representing Restore properties of a Confidential Ledger Resource.
+type Restore struct {
+	// REQUIRED; Fileshare where the ledger backup is stored.
+	FileShareName *string
+
+	// REQUIRED; The region the ledger is being restored to.
+	RestoreRegion *string
+
+	// REQUIRED; SAS URI used to access the backup fileshare.
+	URI *string
+}
+
+// RestoreResponse - Object representing the restore response of a Confidential Ledger Resource.
+type RestoreResponse struct {
+	// READ-ONLY; Response body stating if the ledger is being restored.
+	Message *string
 }
 
 // SystemData - Metadata pertaining to creation and last modification of the resource.