+++ sdk/resourcemanager/confidentialledger/armconfidentialledger/models_serde.go
@@ -50,6 +50,64 @@ func (a *AADBasedSecurityPrincipal) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type Backup.
+func (b Backup) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "restoreRegion", b.RestoreRegion)
+	populate(objectMap, "uri", b.URI)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type Backup.
+func (b *Backup) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", b, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "restoreRegion":
+			err = unpopulate(val, "RestoreRegion", &b.RestoreRegion)
+			delete(rawMsg, key)
+		case "uri":
+			err = unpopulate(val, "URI", &b.URI)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", b, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type BackupResponse.
+func (b BackupResponse) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "message", b.Message)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type BackupResponse.
+func (b *BackupResponse) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", b, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "message":
+			err = unpopulate(val, "Message", &b.Message)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", b, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type CertBasedSecurityPrincipal.
 func (c CertBasedSecurityPrincipal) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -198,6 +256,41 @@ func (c *ConfidentialLedger) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type CustomizationData.
+func (c CustomizationData) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "hostLevel", c.HostLevel)
+	populate(objectMap, "maxBodySize", c.MaxBodySize)
+	populate(objectMap, "subjectName", c.SubjectName)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type CustomizationData.
+func (c *CustomizationData) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", c, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "hostLevel":
+			err = unpopulate(val, "HostLevel", &c.HostLevel)
+			delete(rawMsg, key)
+		case "maxBodySize":
+			err = unpopulate(val, "MaxBodySize", &c.MaxBodySize)
+			delete(rawMsg, key)
+		case "subjectName":
+			err = unpopulate(val, "SubjectName", &c.SubjectName)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", c, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type DeploymentType.
 func (d DeploymentType) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -234,9 +327,11 @@ func (l LedgerProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "aadBasedSecurityPrincipals", l.AADBasedSecurityPrincipals)
 	populate(objectMap, "certBasedSecurityPrincipals", l.CertBasedSecurityPrincipals)
+	populate(objectMap, "enclavePlatform", l.EnclavePlatform)
 	populate(objectMap, "identityServiceUri", l.IdentityServiceURI)
 	populate(objectMap, "ledgerInternalNamespace", l.LedgerInternalNamespace)
 	populate(objectMap, "ledgerName", l.LedgerName)
+	populate(objectMap, "ledgerSku", l.LedgerSKU)
 	populate(objectMap, "ledgerType", l.LedgerType)
 	populate(objectMap, "ledgerUri", l.LedgerURI)
 	populate(objectMap, "provisioningState", l.ProvisioningState)
@@ -259,6 +354,9 @@ func (l *LedgerProperties) UnmarshalJSON(data []byte) error {
 		case "certBasedSecurityPrincipals":
 			err = unpopulate(val, "CertBasedSecurityPrincipals", &l.CertBasedSecurityPrincipals)
 			delete(rawMsg, key)
+		case "enclavePlatform":
+			err = unpopulate(val, "EnclavePlatform", &l.EnclavePlatform)
+			delete(rawMsg, key)
 		case "identityServiceUri":
 			err = unpopulate(val, "IdentityServiceURI", &l.IdentityServiceURI)
 			delete(rawMsg, key)
@@ -267,6 +365,9 @@ func (l *LedgerProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "ledgerName":
 			err = unpopulate(val, "LedgerName", &l.LedgerName)
+			delete(rawMsg, key)
+		case "ledgerSku":
+			err = unpopulate(val, "LedgerSKU", &l.LedgerSKU)
 			delete(rawMsg, key)
 		case "ledgerType":
 			err = unpopulate(val, "LedgerType", &l.LedgerType)
@@ -370,6 +471,64 @@ func (m *ManagedCCF) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type ManagedCCFBackup.
+func (m ManagedCCFBackup) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "restoreRegion", m.RestoreRegion)
+	populate(objectMap, "uri", m.URI)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type ManagedCCFBackup.
+func (m *ManagedCCFBackup) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", m, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "restoreRegion":
+			err = unpopulate(val, "RestoreRegion", &m.RestoreRegion)
+			delete(rawMsg, key)
+		case "uri":
+			err = unpopulate(val, "URI", &m.URI)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", m, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type ManagedCCFBackupResponse.
+func (m ManagedCCFBackupResponse) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "message", m.Message)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type ManagedCCFBackupResponse.
+func (m *ManagedCCFBackupResponse) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", m, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "message":
+			err = unpopulate(val, "Message", &m.Message)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", m, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type ManagedCCFList.
 func (m ManagedCCFList) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -406,11 +565,14 @@ func (m ManagedCCFProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "appName", m.AppName)
 	populate(objectMap, "appUri", m.AppURI)
+	populate(objectMap, "customizationData", m.CustomizationData)
 	populate(objectMap, "deploymentType", m.DeploymentType)
+	populate(objectMap, "enclavePlatform", m.EnclavePlatform)
 	populate(objectMap, "identityServiceUri", m.IdentityServiceURI)
 	populate(objectMap, "memberIdentityCertificates", m.MemberIdentityCertificates)
 	populate(objectMap, "nodeCount", m.NodeCount)
 	populate(objectMap, "provisioningState", m.ProvisioningState)
+	populate(objectMap, "runningState", m.RunningState)
 	return json.Marshal(objectMap)
 }
 
@@ -429,8 +591,14 @@ func (m *ManagedCCFProperties) UnmarshalJSON(data []byte) error {
 		case "appUri":
 			err = unpopulate(val, "AppURI", &m.AppURI)
 			delete(rawMsg, key)
+		case "customizationData":
+			err = unpopulate(val, "CustomizationData", &m.CustomizationData)
+			delete(rawMsg, key)
 		case "deploymentType":
 			err = unpopulate(val, "DeploymentType", &m.DeploymentType)
+			delete(rawMsg, key)
+		case "enclavePlatform":
+			err = unpopulate(val, "EnclavePlatform", &m.EnclavePlatform)
 			delete(rawMsg, key)
 		case "identityServiceUri":
 			err = unpopulate(val, "IdentityServiceURI", &m.IdentityServiceURI)
@@ -443,6 +611,71 @@ func (m *ManagedCCFProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "provisioningState":
 			err = unpopulate(val, "ProvisioningState", &m.ProvisioningState)
+			delete(rawMsg, key)
+		case "runningState":
+			err = unpopulate(val, "RunningState", &m.RunningState)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", m, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type ManagedCCFRestore.
+func (m ManagedCCFRestore) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "fileShareName", m.FileShareName)
+	populate(objectMap, "restoreRegion", m.RestoreRegion)
+	populate(objectMap, "uri", m.URI)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type ManagedCCFRestore.
+func (m *ManagedCCFRestore) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", m, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "fileShareName":
+			err = unpopulate(val, "FileShareName", &m.FileShareName)
+			delete(rawMsg, key)
+		case "restoreRegion":
+			err = unpopulate(val, "RestoreRegion", &m.RestoreRegion)
+			delete(rawMsg, key)
+		case "uri":
+			err = unpopulate(val, "URI", &m.URI)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", m, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type ManagedCCFRestoreResponse.
+func (m ManagedCCFRestoreResponse) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "message", m.Message)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type ManagedCCFRestoreResponse.
+func (m *ManagedCCFRestoreResponse) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", m, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "message":
+			err = unpopulate(val, "Message", &m.Message)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -592,6 +825,68 @@ func (r *ResourceProviderOperationList) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type Restore.
+func (r Restore) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "fileShareName", r.FileShareName)
+	populate(objectMap, "restoreRegion", r.RestoreRegion)
+	populate(objectMap, "uri", r.URI)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type Restore.
+func (r *Restore) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", r, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "fileShareName":
+			err = unpopulate(val, "FileShareName", &r.FileShareName)
+			delete(rawMsg, key)
+		case "restoreRegion":
+			err = unpopulate(val, "RestoreRegion", &r.RestoreRegion)
+			delete(rawMsg, key)
+		case "uri":
+			err = unpopulate(val, "URI", &r.URI)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", r, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type RestoreResponse.
+func (r RestoreResponse) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "message", r.Message)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type RestoreResponse.
+func (r *RestoreResponse) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", r, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "message":
+			err = unpopulate(val, "Message", &r.Message)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", r, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type SystemData.
 func (s SystemData) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -660,7 +955,7 @@ func populateAny(m map[string]any, k string, v any) {
 }
 
 func unpopulate(data json.RawMessage, fn string, v any) error {
-	if data == nil {
+	if data == nil || string(data) == "null" {
 		return nil
 	}
 	if err := json.Unmarshal(data, v); err != nil {